requirements.txt
pydantic
python-dotenv
requests
playwright>=1.40
tqdm
